--- a/config.default/nginx.conf
+++ b/config.default/nginx.conf
@@ -46,6 +46,8 @@ http {
   resolver 8.8.8.8;
   charset UTF-8;
 
+  lua_shared_dict cache 10m;
+
   init_by_lua "kong = require 'kong'; kong.init()";
 
   server {