--- a/config.default/nginx.conf
+++ b/config.default/nginx.conf
@@ -46,7 +46,7 @@ http {
   resolver 8.8.8.8;
   charset UTF-8;
 
-  lua_shared_dict cache 10m;
+  lua_shared_dict cache 512m;
 
   init_by_lua "kong = require 'kong'; kong.init()";