--- a/spec/unit/utils_spec.lua
+++ b/spec/unit/utils_spec.lua
@@ -3,6 +3,15 @@ local cjson = require "cjson"
 
 describe("Utils #utils", function()
 
+  describe("Cache", function()
+    it("should return a valid API cache key", function()
+      assert.are.equal("apis/httpbin.org", utils.cache_api_key("httpbin.org"))
+    end)
+    it("should return a valid PLUGIN cache key", function()
+      assert.are.equal("plugins/authentication/api123/app123", utils.cache_plugin_key("authentication", "api123", "app123"))
+    end)
+  end)
+
   describe("HTTP", function()
     describe("GET", function()
       it("should return a valid GET response", function()