--- a/src/constants.lua
+++ b/src/constants.lua
@@ -16,6 +16,10 @@ return {
     RATELIMIT_LIMIT = "X-RateLimit-Limit",
     RATELIMIT_REMAINING = "X-RateLimit-Remaining"
   },
+  CACHE = {
+    APIS = "apis",
+    PLUGINS = "plugins"
+  },
   AUTHENTICATION = {
     QUERY = "query",
     BASIC = "basic",