--- a/src/kong/core/access.lua
+++ b/src/kong/core/access.lua
@@ -1,4 +1,5 @@
 local stringy = require "stringy"
+local constants = require "kong.constants"
 
 local _M = {}
 
@@ -21,19 +22,7 @@ local function skip_authentication(headers)
   return headers["expect"] and _M.starts_with(headers["expect"], "100")
 end
 
-function _M.execute(conf)
-  -- Retrieving the API from the Host that has been requested
-
-  local host = stringy.split(ngx.var.http_host, ":")[1]
-
-  local cache_value = utils.cache_get("wot")
-  if not cache_value then
-    print("No cache")
-    utils.cache_save("wot", "ciao", 10)
-  else
-    print("Cached: "..cache_value)
-  end
-
+local function load_api(host)
   local apis, err = dao.apis:find_by_keys({public_dns = host})
   if err then
     ngx.log(ngx.ERR, err.message)
@@ -41,13 +30,26 @@ function _M.execute(conf)
   elseif not apis or #apis == 0 then
     utils.not_found("API not found")
   end
+  return apis[1]
+end
 
-  local api = apis[1]
+function _M.execute(conf)
+  -- Retrieving the API from the Host that has been requested
+  local host = stringy.strip(stringy.split(ngx.var.http_host, ":")[1])
+
+  local cache_key = utils.cache_api_key(host)
+  local api = utils.cache_get(cache_key)
+  if not api then
+    api = load_api(host)
+    local ok, err = utils.cache_set(cache_key, api, 5)
+    if not ok then
+      ngx.log(ngx.ERR, err)
+    end
+  end
 
   -- Setting the backend URL for the proxy_pass directive
   ngx.var.backend_url = get_backend_url(api) .. ngx.var.request_uri
 
-  -- TODO: Move this away from here
   -- There are some requests whose authentication needs to be skipped
   if skip_authentication(ngx.req.get_headers()) then
     return -- Returning and keeping the Lua code running to the next handler