This pull request is the first step in implementing #15, with the final step being having an invalidation system in place.

It implements a time-expiration caching system configurable through `kong.yaml` configuration file.

```yaml
# Cache configuration
cache:
  expiration: 5 # In seconds
```

I'm creating the pull request without merging, and will post the results of the benchmark below.