@thefosk 2015-02-28
# Benchmarks

The benchmarks have been executed on the EC2 environment, using:

```
siege -c 300 -t 60S ...
```

### Without cache (version `0.0.1-beta`)

```
Transactions:          70626 hits
Availability:          98.53 %
Elapsed time:          21.80 secs
Data transferred:         0.88 MB
Response time:            0.09 secs
Transaction rate:      3239.72 trans/sec
Throughput:           0.04 MB/sec
Concurrency:          297.06
Successful transactions:       70626
Failed transactions:          1052
Longest transaction:          1.37
Shortest transaction:         0.01
```

### With cache (default=5s)

```
Transactions:         535318 hits
Availability:          99.77 %
Elapsed time:          43.45 secs
Data transferred:         6.64 MB
Response time:            0.02 secs
Transaction rate:     12320.32 trans/sec
Throughput:           0.15 MB/sec
Concurrency:          295.57
Successful transactions:      535318
Failed transactions:          1209
Longest transaction:          5.02
Shortest transaction:         0.00
```