@thibaultCha 2015-08-03
This is in Kong since the first release.