@thefosk 2015-08-03
@samek bear in mind that is the internal cache of Kong for its internal entities, not an HTTP caching layer for the upstream APIs.