@samek 2015-08-03
@thefosk  exactly what I thought it was for (http upstream). :( 

Any plans for that ?