@mashapedeployment 2015-08-04
@samek https://github.com/Mashape/kong/issues?q=caching+is%3Aissue