@samek 2015-08-04
ok in the mean time I've added it to the nginx config in kong.yml  - for docker. 
If someone want's to try (it'll be caching all apis's). 

Add parts between the ##CUSTOM CACHING##

     --- a/config.docker/kong.yml
     +++ b/config.docker/kong.yml
     @@ -35,7 +35,8 @@ database: cassandra
      databases_available:
        cassandra:
          properties:
     -      hosts: "cassandra"
     +      hosts: 
     +CASSANDRAHOSTS
            port: 9042
            timeout: 1000
            keyspace: kong
     @@ -124,8 +125,13 @@ nginx: |
          ';
      
          init_worker_by_lua 'kong.exec_plugins_init_worker()';
     +    ##CUSTOM CACHING##
     +    proxy_cache_path  /tmp/ levels=1 keys_zone=apicache:10m;
     +    ##CUSTOM CACHING##
 
          server {
     +      set_real_ip_from 0.0.0.0/0;
     +      real_ip_header Cres-Client-IP;
            server_name _;
            listen {{proxy_port}};
            listen {{proxy_ssl_port}} ssl;
     @@ -144,6 +150,15 @@ nginx: |
              # Authenticate the user and load the API info
              access_by_lua 'kong.exec_plugins_access()';
      
     +        ##CUSTOM CACHING##
     +        proxy_cache apicache;
     +        proxy_cache_use_stale updating;
     +        proxy_cache_lock on;
     +        proxy_cache_valid 200 10s;
     +        proxy_ignore_headers X-Accel-Expires Expires Cache-Control;
     +        add_header X-Proxy-Cache $upstream_cache_status;
     +        ##CUSTOM CACHING##
     +
              # Proxy the request
              proxy_set_header X-Real-IP $remote_addr;
              proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
     @@ -203,4 +218,4 @@ nginx: |
            # Do not remove, additional configuration placeholder for some plugins
            # {{additional_configuration}}
          }